Switch to a regular `library` dub targetType
Which prevents issues with LDC (in combination with dpp-generated .d
files) and reggae (which doesn't properly support sourceLibraries).

The net effect is that dub compiles the few .d files here separately
now, *not* inlining them as source files into each project depending
on mupdf-d (and so compiling them with the other modules in one
compiler invocation by default).